+++ java-service-processor/src/main/java/internal/nbbrd/service/definition/ServiceDefinitionChecker.java
@@ -27,7 +27,6 @@ import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.Modifier;
 import javax.lang.model.element.TypeElement;
-import javax.lang.model.type.DeclaredType;
 import javax.lang.model.type.PrimitiveType;
 import javax.lang.model.type.TypeKind;
 import javax.lang.model.type.TypeMirror;
@@ -39,7 +38,6 @@ import java.util.*;
 import java.util.function.Predicate;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
-import java.util.stream.Stream;
 
 import static javax.lang.model.element.Modifier.*;
 
@@ -219,8 +217,8 @@ final class ServiceDefinitionChecker {
     }
 
     private boolean checkBatch(LoadDefinition definition, TypeElement service) {
-        if (definition.getBatchType().isPresent()) {
-            TypeElement x = env.asTypeElement(definition.getBatchType().get());
+        if (definition.getBatch().isPresent()) {
+            TypeElement x = env.asTypeElement(definition.getBatch().get().getType());
             if (x.getKind() != ElementKind.INTERFACE && !x.getModifiers().contains(ABSTRACT)) {
                 env.error(service, "[RULE_B1] Batch type must be an interface or an abstract class");
                 return false;
@@ -234,13 +232,15 @@ final class ServiceDefinitionChecker {
     }
 
     private Predicate<ExecutableElement> batchMethodFilter(TypeElement service) {
-        DeclaredType streamType = env.getTypeUtils().getDeclaredType(env.asTypeElement(Stream.class), service.asType());
         return method -> method.getModifiers().contains(PUBLIC)
                 && !method.getModifiers().contains(STATIC)
                 && method.getParameters().isEmpty()
-                && method.getSimpleName().contentEquals("getProviders")
-                && env.getTypeUtils().isAssignable(method.getReturnType(), streamType)
+                && isBatchReturnType(method.getReturnType(), service)
                 && !hasCheckedExceptions(method);
+    }
+
+    private boolean isBatchReturnType(TypeMirror returnType, TypeElement service) {
+        return BatchDefinition.MethodReturnKind.resolve(returnType, service, env.getTypeUtils(), env.getElementUtils()) != null;
     }
 
     private boolean checkInstanceFactories(TypeElement annotatedElement, TypeMirror type, TypeInstantiator instance) {