+++ java-service-processor/src/test/resources/definition/TestBatchCollectionReturnType.java
@@ -1,0 +1,14 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+
+import java.util.Collection;
+
+@ServiceDefinition(batchType = TestBatchCollectionReturnType.SomeBatch.class)
+interface TestBatchCollectionReturnType {
+
+    interface SomeBatch {
+
+        Collection<TestBatchCollectionReturnType> getProviders();
+    }
+}