+++ java-service-processor/src/test/resources/definition/TestBatchCustomMethodName.java
@@ -1,0 +1,15 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+
+import java.util.stream.Stream;
+
+@ServiceDefinition(batchType = TestBatchCustomMethodName.SomeBatch.class)
+interface TestBatchCustomMethodName {
+
+    interface SomeBatch {
+
+        Stream<TestBatchCustomMethodName> getAll();
+    }
+}
+